Created utility methods to find all properties of type in process
