+++ Source/TextToSpeech-Component/Editor/TextToSpeechEditorUtils.cs
@@ -65,103 +65,5 @@ namespace VRBuilder.Editor.TextToSpeech
                 }
             }
         }
-
-        public static IEnumerable<TextToSpeechAudio> GetTTSFromProcess(IProcess process)
-        {
-            List<TextToSpeechAudio> tts = new List<TextToSpeechAudio>();
-
-            foreach (IChapter chapter in process.Data.GetChildren())
-            {
-                tts.AddRange(GetTTSFromChapter(chapter));
-            }
-
-            return tts;
-        }
-
-        public static IEnumerable<TextToSpeechAudio> GetTTSFromChapter(IChapter chapter)
-        {
-            List<TextToSpeechAudio> tts = new List<TextToSpeechAudio>();
-
-            foreach (IStep step in chapter.Data.GetChildren())
-            {
-                tts.AddRange(GetTTSFromStep(step));
-            }
-
-            return tts;
-        }
-
-        public static IEnumerable<TextToSpeechAudio> GetTTSFromStep(IStep step)
-        {
-            List<TextToSpeechAudio> tts = new List<TextToSpeechAudio>();
-
-            IEnumerable<IBehaviorCollection> behaviorCollections = step.Data.GetChildren().Where(child => child is BehaviorCollection).Cast<IBehaviorCollection>();
-
-            foreach (IBehaviorCollection behaviorCollection in behaviorCollections)
-            {
-                foreach (IBehavior behavior in behaviorCollection.Data.Behaviors)
-                {
-                    tts.AddRange(GetTTSFromBehavior(behavior));
-
-                    if (behavior.Data is IEntityCollectionData<IBehavior>)
-                    {
-                        IEntityCollectionData<IBehavior> data = behavior.Data as IEntityCollectionData<IBehavior>;
-
-                        foreach (IBehavior childBehavior in data.GetChildren())
-                        {
-                            tts.AddRange(GetTTSFromBehavior(childBehavior));
-                        }
-                    }
-
-                    if (behavior.Data is IEntityCollectionData<IChapter>)
-                    {
-                        IEntityCollectionData<IChapter> data = behavior.Data as IEntityCollectionData<IChapter>;
-
-                        foreach (IChapter childChapter in data.GetChildren())
-                        {
-                            tts.AddRange(GetTTSFromChapter(childChapter));
-                        }
-                    }
-                }
-
-            }
-
-            return tts;
-        }
-
-        public static IEnumerable<TextToSpeechAudio> GetTTSFromBehavior(IBehavior behavior)
-        {
-            List<TextToSpeechAudio> tts = new List<TextToSpeechAudio>();
-
-            IEnumerable<PropertyInfo> properties = behavior.Data.GetType().GetProperties();
-
-            foreach (PropertyInfo property in properties)
-            {
-                IEnumerable<ParameterInfo> indexes = property.GetIndexParameters();
-
-                if (indexes.Count() > 0)
-                {
-                    foreach (ParameterInfo index in indexes)
-                    {
-                        object value = property.GetValue(behavior.Data, new object[] { index });
-
-                        if (value != null && (value.GetType() == typeof(TextToSpeechAudio) || value.GetType().IsSubclassOf(typeof(TextToSpeechAudio))))
-                        {
-                            tts.Add(value as TextToSpeechAudio);
-                        }
-                    }
-                }
-                else
-                {
-                    object value = property.GetValue(behavior.Data);
-
-                    if (value != null && (value.GetType() == typeof(TextToSpeechAudio) || value.GetType().IsSubclassOf(typeof(TextToSpeechAudio))))
-                    {
-                        tts.Add(value as TextToSpeechAudio);
-                    }
-                }
-            }
-
-            return tts;
-        }
     }
 }
